make type conversion more explicit.

+++ src/compute_temp.cpp
@@ -64,7 +64,7 @@ void ComputeTemp::dof_compute()
 {
   double natoms = group->count(igroup);
   dof = domain->dimension * natoms;
-  dof -= extra_dof + fix_dof;
+  dof -= static_cast<double>(extra_dof + fix_dof);
   if (dof > 0.0) tfactor = force->mvv2e / (dof * force->boltz);
   else tfactor = 0.0;
 }
